+++ src/contains.js
@@ -1,10 +1,8 @@
 // @flow
 
-/**
- * Check if a string or array contains an element
- */
-
-const contains = (ls: string | any[], el: any): bool =>
-  ls.includes(el)
+const contains = (ls: string | any[], el: any): bool => {
+  console.error('`contains` is deprecated. Please use `elem` instead.')
+  return ls.includes(el)
+}
 
 export default contains