+++ src/curry2.js
@@ -1,13 +1,5 @@
-/**
- * Takes a function and returns a function that takes two arguments
- * Produces a curried function
- * @param {function} f
- * @param {any} a
- * @param {any} b
- * @returns {function}
- */
-
 function curry2 (f) {
+  console.error('`curry2` is deprecated. Please use `curry` instead.')
   function curried (a, b) {
     switch (arguments.length) {
       case 0: return curried