+++ src/isInteger.js
@@ -1,10 +1,7 @@
 // @flow
 
-/**
- * Returns true if the value is an integer
- */
-
 const isInteger = Number.isInteger || function (n: number): bool {
+  console.error('`isInteger` is depreacted. Please use `Number.isInteger` instead.')
   return (n | 0) === n
 }