+++ src/minifyString.js
@@ -4,7 +4,9 @@
  * Roughly minify a string
  */
 
-const minifyString = (str: string): string =>
-  str.replace(/\n/g, '').replace(/\s\s+/g, ' ')
+const minifyString = (str: string): string => {
+  console.error('`minifyString` is depreacted. Please use `collapseWhitespace` instead.')
+  return str.replace(/\n/g, '').replace(/\s\s+/g, ' ')
+}
 
 export default minifyString