+++ src/notElem.js
@@ -3,8 +3,8 @@
 import elem from './elem'
 
 /**
- * The inverse of `elem`, here because of Haskell's `notElem`
- * Like `!contains`, with reversed args
+ * The opposite of `elem`
+ * Returns true if the element is not in the string or array
  */
 
 const notElem = (el: any, ls: string | any[]): bool =>