+++ src/objectAssign.js
@@ -1,11 +1,7 @@
 // @flow
 
-/**
- * `Object.assign`
- * Copy vals of all enumerable own properties from source obj to target obj
- */
-
 const objectAssign = Object.assign || function (target: Object, source: Object): Object {
+  console.error('`objectAssign` is deprecated. Please use `Object.assign` instead.')
   for (let key in source) {
     if (source.hasOwnProperty(key)) {
       target[key] = source[key]