+++ src/toMap.js
@@ -5,6 +5,7 @@
  */
 
 const toMap = (arr: any[]): Object => {
+  console.error('`toMap` is deprecated. Please use native `Map` instead.')
   const m = {}
   arr.forEach((e, i) => {
     e.index = i