+++ src/toType.js
@@ -4,7 +4,9 @@
  * @returns {string}
  */
 
-const toType = (val) =>
-  ({}).toString.call(val).match(/\s([a-zA-Z]+)/)[1].toLowerCase()
+const toType = (val) => {
+  console.error('`toType` is depreacted. Use `typeOf` instead.')
+  return ({}).toString.call(val).match(/\s([a-zA-Z]+)/)[1].toLowerCase()
+}
 
 export default toType