+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+* `0.8.3` &mdash; **2017-05-25** &mdash; *important* &mdash; deprecations
+  * `contains` (use `elem`)
+  * `curry2`, `curry3`, `curry4` (use `curry`)
+  * `head` (use `car`)
+  * `imul` (use `Math.imul`)
+  * `isInteger` (use `Number.isInteger`)
+  * `keys` (use `Object.keys`)
+  * `moustache` (use template literals)
+  * `newlinesToSpaces` (use `collapseWhitespace`)
+  * `notContains` (use `notElem`)
+  * `objectAssign` (use `Object.assign`)
+  * `objectClone` (use `clone`)
+  * `tail` (use `cdr`)
 * `0.8.0` &mdash; **2017-05-24** &mdash; *breaking* &mdash; several changes
   * moved full documentation to [this file](./DOC.md)
   * added `isSet`